Optimize avro-name-p
The auxiliary functions were not being inlined before so the declaims
were added, and the functions were reordered. Type declarations also
use fixnum now instead of integer and simple-string instead of string,
and schar is utilized instead of the more generic char.

The type declaration for the input name was also removed since this
function is supposed to work for any input type.